refactor: update `DtoUserIU` and `User` entity fields
- Removed `@Enumerated(EnumType.STRING)` annotation from `DtoUserIU`'s `userType` property.
- Changed `departmentId` type from `int` to `Short` in the `User` entity for better alignment with database constraints.

+++ src/main/java/com/mrtkyr/classqroom/entity/User.java
@@ -41,7 +41,7 @@ public class User {
 
     //TODO When the Department entity is added, it will be updated.
     @Column(name = "department_id", nullable = false)
-    private int departmentId;
+    private Short departmentId;
 
     @Column(name = "is_in_course", nullable = false)
     private boolean inCourse;
